+++ src/main/java/org/plumelib/util/ArraysPlume.java
@@ -409,8 +409,6 @@ public final class ArraysPlume {
     if (a.length == 0) {
       throw new ArrayIndexOutOfBoundsException("Empty array passed to elementRange(int[])");
     }
-    @SuppressWarnings(
-        "allcheckers:purity.not.deterministic.call") // use deterministic parts of object
     int[] minAndMax = minAndMax(a);
     return minAndMax[1] - minAndMax[0];
   }
@@ -431,8 +429,6 @@ public final class ArraysPlume {
     if (a.length == 0) {
       throw new ArrayIndexOutOfBoundsException("Empty array passed to elementRange(long[])");
     }
-    @SuppressWarnings(
-        "allcheckers:purity.not.deterministic.call") // use deterministic parts of object
     long[] minAndMax = minAndMax(a);
     return minAndMax[1] - minAndMax[0];
   }
@@ -2194,8 +2190,6 @@ public final class ArraysPlume {
    * @return a string representation of the array
    * @throws IllegalArgumentException if a is not an array
    */
-  @SuppressWarnings(
-      "allcheckers:purity") // defensive coding: throw exception when argument is invalid
   @SideEffectFree
   public static @NonDet String toString(Object a) {
     if (a == null) {
@@ -2234,8 +2228,6 @@ public final class ArraysPlume {
    * @return the length of the array
    * @throws IllegalArgumentException if obj is null or is not an array
    */
-  @SuppressWarnings(
-      "allcheckers:purity") // defensive coding: throw exception when argument is invalid
   @Pure
   public static @NonNegative int length(Object a) throws IllegalArgumentException {
     if (a == null) {
@@ -3152,9 +3144,6 @@ public final class ArraysPlume {
      *     argument is less than, equal to, or greater than the second argument
      */
     @Pure
-    @SuppressWarnings(
-        "nullness:override.param.invalid") // CF bug: doesn't expand annotations on array
-    // elements with @Poly
     // The signature on this method is unnecessarily strict because it
     // requires that the component types be identical.  The signature should
     // be compare(@PolyNull(1) String[], @PolyNull(2) String[]), but the
@@ -3205,9 +3194,6 @@ public final class ArraysPlume {
      *     argument is less than, equal to, or greater than the second argument
      */
     @Pure
-    @SuppressWarnings(
-        "nullness:override.param.invalid") // CF bug: doesn't expand annotations on array
-    // elements with @Poly
     // The signature on this method is unnecessarily strict because it
     // requires that the component types be identical.  The signature should
     // be compare(@PolyNull(1) T[], @PolyNull(2) T[]), but the
@@ -3267,9 +3253,6 @@ public final class ArraysPlume {
      *     argument is less than, equal to, or greater than the second argument
      */
     @Pure
-    @SuppressWarnings(
-        "nullness:override.param.invalid") // CF bug: doesn't expand annotations on array
-    // elements with @Poly
     // The signature on this method is unnecessarily strict because it
     // requires that the component types be identical.  The signature should
     // be compare(@PolyNull(1) Object[], @PolyNull(2) Object[]), but the
@@ -3387,9 +3370,6 @@ public final class ArraysPlume {
      *     argument is less than, equal to, or greater than the second argument
      */
     @Pure
-    @SuppressWarnings(
-        "nullness:override.param.invalid") // CF bug: doesn't expand annotations on array
-    // elements with @Poly
     // The signature on this method is unnecessarily strict because it
     // requires that the component types be identical.  The signature should
     // be compare(@PolyNull(1) T[], @PolyNull(2) T[]), but the
@@ -3455,9 +3435,6 @@ public final class ArraysPlume {
      *     argument is less than, equal to, or greater than the second argument
      */
     @Pure
-    @SuppressWarnings(
-        "nullness:override.param.invalid") // CF bug: doesn't expand annotations on array
-    // elements with @Poly
     // The signature on this method is unnecessarily strict because it
     // requires that the component types be identical.  The signature should
     // be compare(@PolyNull(1) Object[], @PolyNull(2) Object[]), but the