+++ src/main/java/org/plumelib/util/UniqueIdMap.java
@@ -9,6 +9,17 @@ import org.checkerframework.checker.determinism.qual.*;
  *
  * <p>The IDs count up from 0, per instance of UniqueIdMap. When you look up an object in this map,
  * it is given a unique ID if it didn't already have one.
+ *
+ * <p>Typical use:
+ *
+ * <ol>
+ *   <li>Define a field:
+ *       <pre><code>
+ * /** Unique ids for trees. *&#47;
+ * static UniqueIdMap&lt;Tree&gt; treeUids = new UniqueIdMap&lt;&gt;();
+ * </code></pre>
+ *   <li>Wherever you would call {@code x.hashCode()}, instead call {@code treeUids.get(x)}.
+ * </ol>
  */
 public class UniqueIdMap<E> {