+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Plume-Util change log
 
+## 1.2.1 (forthcoming)
+
+- New methods:
+   - `UtilPlume.usedMemory` with no formal parameter
+
+## 1.2.0
+
+- New methods:
+   - `UtilPlume.toStringAndClass`
+   - `UtilPlume.mapToStringAndClass`
+   - `UtilPlume.gc`
+   - `UtilPlume.usedMemory`
+- Remove field `UniqueId.nextUid`
+
 ## 1.1.7
 
 - New class `UniqueIdMap`